remove color pkg and add progress bar

+++ Go/sort.go
@@ -7,8 +7,9 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
+	"time"
 
-	"github.com/fatih/color"
+	"github.com/schollz/progressbar/v3"
 )
 
 type Data struct {
@@ -28,18 +29,19 @@ var (
 	torrent  = Data{"Torrent", "torrent"}
 	exe      = Data{"Exe", "exe"}
 	allData  = []Data{images, videos, music, document, psd, pdf, archive, torrent, exe}
-	// EMPTY FOLDERS
-	emptyFolders int
+	// GET ALL FOLDERS
+	folders int
+	// GET ALL SORTED FILES
+	files int64
 )
-
-// GLOBAL FOLDER
-const sortedFiles = "Sorted Files"
 
 func sorting() bool {
 	// IF FILE TO SORT EXIST = true, DEFAULT = false
 	fileToSortExists := false
 	// CHECK ALL PATH
 	err := filepath.Walk(".", func(path string, info os.FileInfo, err error) error {
+		// GLOBAL FOLDER
+		const sortedFiles = "Sorted Files"
 		// SKIP SORTED FILES DIR
 		SkipDir := fs.SkipDir
 		if path == sortedFiles {
@@ -48,12 +50,10 @@ func sorting() bool {
 		if err != nil {
 			fmt.Println(err)
 		}
-		// CALCULATE EMPTY FOLDERS
+		// CALCULATE FOLDERS
 		if info.IsDir() {
-			emptyFolders = emptyFolders + 1
+			folders = folders + 1
 		}
-		// GET MODIFICATION FILE TIME
-		modTimeFolder := info.ModTime().Format("2006")
 		// CHECK FILES
 		for _, data := range allData {
 			extensions := strings.Split(data.extension, " ")
@@ -61,22 +61,24 @@ func sorting() bool {
 				fileExtname := strings.Trim(filepath.Ext(info.Name()), ".")
 				// SORTING
 				if info.Name() != "sort_windows.exe" && info.Name() != "sort.exe" && strings.ToLower(fileExtname) == extension {
+					// CALCULATE FILES
+					if !info.IsDir() {
+						files = files + 1
+					}
+					// GET MODIFICATION FILE TIME
+					modTimeFolder := info.ModTime().Format("2006")
 					// IF FOLDERS NOT EXIST CREATE
 					if _, err := os.Stat(sortedFiles); os.IsNotExist(err) {
 						os.Mkdir(sortedFiles, 0755)
-						color.Yellow("Create " + sortedFiles + " folder")
 					}
 					if _, err := os.Stat(filepath.Join(sortedFiles, modTimeFolder)); os.IsNotExist(err) {
 						os.MkdirAll(filepath.Join(sortedFiles, modTimeFolder), 0755)
-						color.Yellow("Create " + sortedFiles + "/" + modTimeFolder + " folder")
 					}
 					if _, err := os.Stat(filepath.Join(sortedFiles, modTimeFolder, data.folder)); os.IsNotExist(err) {
 						os.Mkdir(filepath.Join(sortedFiles, modTimeFolder, data.folder), 0755)
-						color.Yellow("Create " + sortedFiles + "/" + modTimeFolder + "/" + data.folder + " folder")
 					}
 					// MOVE FILE
 					os.Rename(path, filepath.Join(sortedFiles, modTimeFolder, data.folder, info.Name()))
-					color.Green(path + " moved >> " + filepath.Join(sortedFiles, modTimeFolder, data.folder, info.Name()))
 					fileToSortExists = true
 				}
 			}
@@ -101,27 +103,35 @@ func removeDir(path string, info os.FileInfo) {
 	if err != nil {
 		panic(err)
 	}
-	color.Red(path + " folder removed!!!")
 }
 
 func main() {
 	if sorting() {
-		for i := 0; i < emptyFolders-1; i++ {
-			err := filepath.Walk(".", func(path string, info os.FileInfo, err error) error {
-				if info.IsDir() {
-					removeDir(path, info)
+		fmt.Println("SORTING!!!")
+		bar := progressbar.Default(files)
+		for i := files; i > 0; i-- {
+			bar.Add(1)
+			time.Sleep(2 * time.Millisecond)
+		}
+		if folders-1 > 0 {
+			for i := 0; i < folders-1; i++ {
+				err := filepath.Walk(".", func(path string, info os.FileInfo, err error) error {
+					if info.IsDir() {
+						removeDir(path, info)
+					}
+					return nil
+				})
+				if err != nil {
+					fmt.Println(err)
 				}
-				return nil
-			})
-			if err != nil {
-				fmt.Println(err)
 			}
+			fmt.Println("EMPTY FOLDERS DELETED!!!")
 		}
 	} else {
-		color.Cyan("NOTHING TO SORT!!!")
+		fmt.Println("NOTHING TO SORT!!!")
 	}
 	// UNCOMMENT IF BUILD FOR WINDOWS
-	// var input string
-	// color.Cyan("PRESS ENTER TO CLOSE!!!")
-	// fmt.Scanln(&input)
+	var input string
+	fmt.Println("PRESS ENTER TO CLOSE!!!")
+	fmt.Scanln(&input)
 }
